requirements.txt
asttokens==3.0.1
bleach==6.3.0
certifi==2026.2.25
charset-normalizer==3.4.4
comm==0.2.3
debugpy==1.8.20
decorator==5.2.1
executing==2.2.1
idna==3.11
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.8.0
jupyter_core==5.9.1
kaggle==2.0.0
kagglesdk==0.1.15
matplotlib-inline==0.2.1
nest-asyncio==1.6.0
numpy==2.4.2
packaging==26.0
pandas==3.0.1
parso==0.8.6
pexpect==4.9.0
platformdirs==4.9.2
prompt_toolkit==3.0.52
protobuf==6.33.5
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.3
pyzmq==27.1.0
requests==2.32.5
six==1.17.0
stack-data==0.6.3
text-unidecode==1.3
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
urllib3==2.6.3
wcwidth==0.6.0
webencodings==0.5.1
